Move classes out of test case.
