+++ osu.Desktop.VisualTests/Tests/TestCaseManiaPlayfield.cs
@@ -21,6 +21,7 @@ using OpenTK.Input;
 using osu.Game.Rulesets.UI;
 using osu.Framework.Allocation;
 using osu.Framework.Extensions.IEnumerableExtensions;
+using osu.Game.Rulesets.Mania.UI;
 
 namespace osu.Desktop.VisualTests.Tests
 {
@@ -70,264 +71,6 @@ namespace osu.Desktop.VisualTests.Tests
             {
                 Key = column.Key
             });
-        }
-    }
-
-    public class ManiaPlayfield : Container
-    {
-        public readonly FlowContainer<Column> Columns;
-
-        public ManiaPlayfield(int columnCount)
-        {
-            if (columnCount > 9)
-                throw new ArgumentException($@"{columnCount} columns is not supported.");
-            if (columnCount <= 0)
-                throw new ArgumentException($@"Can't have zero or fewer columns.");
-
-            RelativeSizeAxes = Axes.Y;
-            AutoSizeAxes = Axes.X;
-
-            Children = new Drawable[]
-            {
-                new Box
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Colour = Color4.Black
-                },
-                Columns = new FillFlowContainer<Column>
-                {
-                    RelativeSizeAxes = Axes.Y,
-                    AutoSizeAxes = Axes.X,
-                    Direction = FillDirection.Horizontal,
-                    Padding = new MarginPadding { Left = 1, Right = 1 },
-                    Spacing = new Vector2(1, 0)
-                }
-            };
-
-            for (int i = 0; i < columnCount; i++)
-                Columns.Add(new Column());
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
-        {
-            var columnColours = new Color4[]
-            {
-                colours.RedDark,
-                colours.GreenDark,
-                colours.BlueDark // Special column
-            };
-
-            int columnCount = Columns.Children.Count();
-            int halfColumns = columnCount / 2;
-
-            var keys = new Key[] { Key.A, Key.S, Key.D, Key.F, Key.Space, Key.J, Key.K, Key.L, Key.Semicolon };
-
-            for (int i = 0; i < halfColumns; i++)
-            {
-                Column leftColumn = Columns.Children.ElementAt(i);
-                Column rightColumn = Columns.Children.ElementAt(columnCount - 1 - i);
-
-                Color4 accent = columnColours[i % 2];
-                leftColumn.AccentColour = rightColumn.AccentColour = accent;
-                leftColumn.Key = keys[keys.Length / 2 - halfColumns + i];
-                rightColumn.Key = keys[keys.Length / 2 + halfColumns - i];
-            }
-
-            bool hasSpecial = halfColumns * 2 < columnCount;
-            if (hasSpecial)
-            {
-                Column specialColumn = Columns.Children.ElementAt(halfColumns);
-                specialColumn.IsSpecialColumn = true;
-                specialColumn.AccentColour = columnColours[2];
-                specialColumn.Key = keys[keys.Length / 2];
-            }
-        }
-    }
-
-    public class Column : Container, IHasAccentColour
-    {
-        private const float key_size = 50;
-
-        private const float key_icon_size = 10;
-        private const float key_icon_corner_radius = 3;
-        private const float key_icon_border_radius = 2;
-
-        private const float hit_target_height = 10;
-        private const float hit_target_bar_height = 2;
-
-        private const float column_width = 45;
-        private const float special_column_width = 70;
-
-        private Color4 accentColour;
-        public Color4 AccentColour
-        {
-            get { return accentColour; }
-            set
-            {
-                if (accentColour == value)
-                    return;
-                accentColour = value;
-
-                setAccentColour();
-            }
-        }
-
-        private bool isSpecialColumn;
-        public bool IsSpecialColumn
-        {
-            get { return isSpecialColumn; }
-            set
-            {
-                isSpecialColumn = value;
-                Width = isSpecialColumn ? special_column_width : column_width;
-            }
-        }
-
-        public Key Key;
-
-        private Box background;
-        private Container hitTargetBar;
-        private Container keyIcon;
-
-        public Column()
-        {
-            RelativeSizeAxes = Axes.Y;
-            Width = column_width;
-
-            Children = new Drawable[]
-            {
-                background = new Box
-                {
-                    Name = "Foreground",
-                    RelativeSizeAxes = Axes.Both,
-                    Alpha = 0.2f
-                },
-                new FillFlowContainer
-                {
-                    Name = "Key + hit target",
-                    Anchor = Anchor.BottomCentre,
-                    Origin = Anchor.BottomCentre,
-                    RelativeSizeAxes = Axes.X,
-                    AutoSizeAxes = Axes.Y,
-                    Direction = FillDirection.Vertical,
-                    Children = new[]
-                    {
-                        new Container
-                        {
-                            Name = "Key",
-                            Anchor = Anchor.BottomCentre,
-                            Origin = Anchor.BottomCentre,
-                            RelativeSizeAxes = Axes.X,
-                            Height = key_size,
-                            Children = new Drawable[]
-                            {
-                                new Box
-                                {
-                                    Name = "Key gradient",
-                                    RelativeSizeAxes = Axes.Both,
-                                    ColourInfo = ColourInfo.GradientVertical(Color4.Black, Color4.Black.Opacity(0)),
-                                    Alpha = 0.5f
-                                },
-                                keyIcon = new Container
-                                {
-                                    Name = "Key icon",
-                                    Anchor = Anchor.Centre,
-                                    Origin = Anchor.Centre,
-                                    Size = new Vector2(key_icon_size),
-                                    Masking = true,
-                                    CornerRadius = key_icon_corner_radius,
-                                    BorderThickness = 2,
-                                    BorderColour = Color4.White, // Not true
-                                    Children = new[]
-                                    {
-                                        new Box
-                                        {
-                                            RelativeSizeAxes = Axes.Both,
-                                            Alpha = 0,
-                                            AlwaysPresent = true
-                                        }
-                                    }
-                                }
-                            }
-                        },
-                        new Container
-                        {
-                            Name = "Hit target",
-                            Anchor = Anchor.BottomCentre,
-                            Origin = Anchor.BottomCentre,
-                            RelativeSizeAxes = Axes.X,
-                            Height = hit_target_height,
-                            Children = new Drawable[]
-                            {
-                                new Box
-                                {
-                                    Name = "Background",
-                                    RelativeSizeAxes = Axes.Both,
-                                    Colour = Color4.Black
-                                },
-                                hitTargetBar = new Container
-                                {
-                                    Name = "Bar",
-                                    Anchor = Anchor.BottomCentre,
-                                    Origin = Anchor.BottomCentre,
-                                    RelativeSizeAxes = Axes.X,
-                                    Height = hit_target_bar_height,
-                                    Masking = true,
-                                    Children = new[]
-                                    {
-                                        new Box
-                                        {
-                                            RelativeSizeAxes = Axes.Both
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-
-        private void setAccentColour()
-        {
-            background.Colour = AccentColour;
-
-            hitTargetBar.EdgeEffect = new EdgeEffect
-            {
-                Type = EdgeEffectType.Glow,
-                Radius = 5,
-                Colour = AccentColour.Opacity(0.5f),
-            };
-
-            keyIcon.EdgeEffect = new EdgeEffect
-            {
-                Type = EdgeEffectType.Glow,
-                Radius = 5,
-                Colour = AccentColour.Opacity(0.5f),
-            };
-        }
-
-        protected override bool OnKeyDown(InputState state, KeyDownEventArgs args)
-        {
-            if (args.Key == Key && !args.Repeat)
-            {
-                background.FadeTo(background.Alpha + 0.2f, 50, EasingTypes.OutQuint);
-                keyIcon.ScaleTo(1.4f, 50, EasingTypes.OutQuint);
-            }
-
-            return false;
-        }
-
-        protected override bool OnKeyUp(InputState state, KeyUpEventArgs args)
-        {
-            if (args.Key == Key)
-            {
-                background.FadeTo(0.2f, 800, EasingTypes.OutQuart);
-                keyIcon.ScaleTo(1f, 400, EasingTypes.OutQuart);
-            }
-
-            return false;
         }
     }
 }
