+++ osu.Game.Rulesets.Mania/osu.Game.Rulesets.Mania.csproj
@@ -56,6 +56,7 @@
     <Compile Include="Objects\ManiaBaseHit.cs" />
     <Compile Include="Objects\Note.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UI\Column.cs" />
     <Compile Include="UI\ManiaHitRenderer.cs" />
     <Compile Include="UI\ManiaPlayfield.cs" />
     <Compile Include="ManiaRuleset.cs" />
